requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.49
psycopg2-binary==2.9.7
alembic==1.11.1
pydantic==1.10.9
python-dotenv==1.0.0
requests==2.31.0
